requirements.txt
streamlit
litellm
mysql-connector-python